transport: widen out_list_s::offsetPayload to uint64
The async send queue tracked the payload send-offset (offsetPayload) as
unsigned int. For DB payloads above 4 GiB the offset wrapped, causing the
sender to re-read the tail of the payload from a wrapped (wrong) address.
The receiver got the full byte count but corrupted content above 4 GiB.

This silently corrupted any DataBlock transfer >4 GiB (e.g. the megalarge
convolution-2d coarse DB at 16 GiB), transport-agnostic (TCP and RDMA both
affected). Widening offsetPayload to uint64 restores byte-perfect transfer:
the 16 GiB coarse now arrives with matching send/recv checksum and the
benchmark validates (checksum 3.199790035637e+04, Status: PASS).

+++ libs/src/core/transport/protocol.c
@@ -57,7 +57,7 @@ struct out_list_s {
   unsigned int rank;
   void *payload;
   uint64_t payloadSize;
-  unsigned int offsetPayload;
+  uint64_t offsetPayload;
   void (*free_method)(void *);
 };
